Refactor AB Codec tests.
